MA-24081-1 audio: Fix coverity warning of Unintentional integer overflow
Fix CID 49112035: (#1 of 1): Unintentional integer overflow (OVERFLOW_BEFORE_WIDEN).

Reviewed-on: http://androidsource.nxp.com/project/29826

+++ audio/default/primary/StreamPrimary.cpp
@@ -425,7 +425,7 @@ done:
         return StreamAlsa::refinePosition(position);
     }
     if (!mIsInput) {
-        int64_t buffer_size = mConfig->period_size * mConfig->period_count;
+        int64_t buffer_size = static_cast<int64_t>(mConfig->period_size) * mConfig->period_count;
         position->frames = position->frames > buffer_size ?
             (position->frames - buffer_size) : 0;
     }
